Add files via upload
This PRAAT script takes .wav files and annotates them using PRAAT-specific format called a TextGrid.
The purpose of this script is to detect intervals of silences and sounds and label them accordingly. To do that, PRAAT detects the sounds using a built-in function (some options can be adjusted - see below) and displays all of the created TextGrids so that the user can make manual adjustments/quality control.

WARNING: In case of a high number of files, the processing can take a few minutes and PRAAT will open a lot of windows, so don't be surprised .

The label of the sounding interval is taken from the names of the analyzed .wav  files (example of a name of a .wav file that this script was developed to process: 1_030_wn_5_meisje_WN.wav; however, other names can be used too if the REGEX is adjusted).  

In each file ONLY ONE SOUNDING INTERVAL IS EXPECTED!