Create wav_to_textgrid_wordincsv.praat
This PRAAT script takes .wav files and annotates them using PRAAT-specific format called a TextGrid. The purpose of this script is to detect intervals of silences and sounds and label them accordingly. To do that, PRAAT detects the sounds using a built-in function (some options can be adjusted - see below) and displays all of the created TextGrids so that the user can make manual adjustments/quality control. The label of the sounding interval is taken from an additional .csv file provided by the user.